requirements.txt
python-dotenv
openai
pyupbit
pandas
numpy
requests
matplotlib
mplfinance
streamlit
plotly



